+++ CHANGELOG.md
@@ -10,10 +10,17 @@ Not yet released.
 * [CS-111][] Ping hub on /xapi/hubs operations
 * [CS-62][] Special-case an error when someone wants to POST to /images/save but instead GETs /images/save.
 * [CS-215][] POST /xapi/docker/pull will now return a 404 rather than a 500 when it cannot find the image you are attempting to pull
+* [CS-318][] Containers returned from `GET /xapi/containers` and `GET /xapi/containers/{id}` now include top-level `status` property. This is mostly equal to the status in the most recent history item on the container, except for mapping the docker event statuses to more user-friendly statuses.
+    * `created` -> `Created`
+    * `started` -> `Running`
+    * `die` -> `Done`
+    * `kill` -> `Killed`
+    * `oom` -> `Killed (Out of memory)`
 
 [CS-111]: https://issues.xnat.org/browse/CS-111
 [CS-62]: https://issues.xnat.org/browse/CS-62
 [CS-215]: https://issues.xnat.org/browse/CS-215
+[CS-318]: https://issues.xnat.org/browse/CS-318
 
 ### Bugfixes